[MS] Fixed circular import and composable for opening file
