+++ client/src/components/workspaces/index.ts
@@ -1,8 +1,7 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
 
+export { WORKSPACES_PAGE_DATA_KEY, WorkspaceDefaultData } from '@/components/workspaces/types';
 export {
-  WORKSPACES_PAGE_DATA_KEY,
-  WorkspaceDefaultData,
   compareWorkspaceRoles,
   takeOwnershipOfWorkspace,
   useWorkspaceContextMenu,
@@ -12,5 +11,5 @@ export {
 import WorkspaceFilter from '@/components/workspaces/WorkspaceFilter.vue';
 import WorkspaceRoleTag from '@/components/workspaces/WorkspaceRoleTag.vue';
 
-export type { WorkspacesPageFilters, WorkspacesPageSavedData } from '@/components/workspaces/utils';
+export type { WorkspacesPageFilters, WorkspacesPageSavedData } from '@/components/workspaces/types';
 export { WorkspaceFilter, WorkspaceRoleTag };
